requirements.txt
fastapi
uvicorn
python_dotenv
tortoise-orm
requests
xlwt
asyncpg
aiomysql
aiosqlite

transformers
sentencepiece
torch